Complete MovieDetailActivity without mediaplayer.

+++ data/src/main/java/mobile/app/data/DefaultMovieRepository.kt
@@ -2,5 +2,5 @@ package mobile.app.data
 
 interface DefaultMovieRepository {
     suspend fun getMovies(sortBy: String,title: String): RestResult<MoviesWithGroup>
-    suspend fun getMovieDetail(movieId: Int): RestResult<MovieDetail>
+    suspend fun getMovieDetail(movieId: Long): RestResult<MovieDetail>
 }
